fix: @react-oauth/google と Next.js 15 RSC の非互換を解消
- GoogleOAuthProvider を ClientProviders から除去し、AuthProvider のみに変更
- login/_components/GoogleLoginForm.tsx を新規作成し GoogleOAuthProvider + GoogleLogin を集約
- login/page.tsx で next/dynamic + ssr:false を使い GoogleLoginForm をクライアント専用で読み込む
- next.config.ts の transpilePackages から @react-oauth/google を除去（ssr:false のため不要）
- seedData.ts のフォーマット差分を Biome に合わせて整形

背景: @react-oauth/google が Next.js 15 の RSC webpack モジュールローダーと非互換で、
ClientProviders でアプリ全体をラップすると起動時に
'Cannot read properties of undefined (reading call)' エラーが発生していた。

(by Claude)
